LPS-54791 RSS: Standardize locations for internal setting classes
