+++ modules/apps/rss/rss-web/src/com/liferay/rss/web/settings/internal/RSSPortletInstanceConfigurationBeanDeclaration.java
@@ -12,11 +12,10 @@
  * details.
  */
 
-package com.liferay.rss.web.configuration.internal;
+package com.liferay.rss.web.settings.internal;
 
-import com.liferay.portal.kernel.settings.definition.SettingsIdMapping;
+import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
 import com.liferay.rss.web.configuration.RSSPortletInstanceConfiguration;
-import com.liferay.rss.web.constants.RSSPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -24,16 +23,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Iván Zaera
  */
 @Component
-public class RSSPortletInstanceSettingsIdMapping implements SettingsIdMapping {
+public class RSSPortletInstanceConfigurationBeanDeclaration
+	implements ConfigurationBeanDeclaration {
 
 	@Override
-	public Class<?> getConfigurationBeanClass() {
+	public Class getConfigurationBeanClass() {
 		return RSSPortletInstanceConfiguration.class;
-	}
-
-	@Override
-	public String getSettingsId() {
-		return RSSPortletKeys.RSS;
 	}
 
 }
